requirements.txt
numpy
pandas
ipykernel
openai
streamlit
python-dotenv
requests
selenium
webdriver_manager
bs4
